Move statuses to statuses/
